Performance :runner: enhancements
- App.tsx had become a bit of god object so Main was created to establish the observers and render the main app components as children
- Created Header.tsx which is RxChart heading icon and org name when logged in
- ClientHeader.tsx renders the two dropdown buttons for the active client.
- The global `development` value was removed from global.d.ts and replaced with `process.env.REACT_APP_DEVELOPMENT === 'true'` in every instance it was used.
- getInitialState.ts index.tsx no longer use the `development` global
  - Some components (ClientButton and ClientDobButton) were expecting this as a prop and these have been removed.
- children prop types use `ReactNode | undefined` where previously they were set to `any`
- Observers no longer take arguments and use the global state directly.
- LandingPage "memoized" `checkoutPage` to reduce the number of re-renders
- MedicineCheckoutPage now has props (where it was using `useGlobal()`:
     ```
     drugLogList: DrugLogRecord[]
     activeTabKey: string
     activeClient: ResidentRecord | null
     ```
