+++ src/components/Buttons/ShadowBox.tsx
@@ -1,8 +1,9 @@
+import {ReactNode} from "react";
 import React from "reactn";
 import './../../styles/shadowbox.css'
 
 interface IProps {
-    children: any;
+    children: ReactNode | undefined;
 }
 
 const ShadowBox = (props: IProps) => {
